Add localized birds
I've made this as if they're simply translations of a bird.
Birds have a default scope which includes localizations by
default. This ensures we always have a localized name available.
I'm not sure how this will work out once we add more lists but
it works for now and is quit elegant.