Added country_code to birding session
This will help us calculate life lists, country lists and more.

I also removed the country from location_name. It was only there to
allow the location search to be exactly as Google returns it. Now
we have a separate country we can omit it from the location name.
Most of the time you'll be looking at sessions through the context
of a club or something local and so the country is superfluous.

I also removed location_address as it has no use.

I also renamed start_time to time. It no longer makes sense after
we removed end_time.